# servicerouter: improve performance