instead of pulling all tasks and then pulling all apps from marathon, this pull request changes `get_apps` to pull all apps and tasks at once using `/v2/apps?embed=apps.tasks`

On a marathon with ~50 apps, this reduced the time to regenerate the config after an event from about 30 seconds to less than two (measured as in #2093):

Before:
```
2015-08-26 16:05:34,551 servicerouter DEBUG updating tasks finished, took 43.9271931648 seconds
2015-08-26 16:06:48,061 servicerouter DEBUG updating tasks finished, took 39.2305569649 seconds
2015-08-26 16:07:23,728 servicerouter DEBUG updating tasks finished, took 35.6545169353 seconds
2015-08-26 16:07:56,529 servicerouter DEBUG updating tasks finished, took 32.7971820831 seconds
```

After:
```
2015-08-26 16:36:56,770 servicerouter DEBUG updating tasks finished, took 1.21059083939 seconds
2015-08-26 16:37:19,629 servicerouter DEBUG updating tasks finished, took 1.67731595039 seconds
2015-08-26 16:37:21,033 servicerouter DEBUG updating tasks finished, took 1.39936995506 seconds
2015-08-26 16:37:22,669 servicerouter DEBUG updating tasks finished, took 1.63086915016 seconds
2015-08-26 16:37:23,522 servicerouter DEBUG updating tasks finished, took 0.84465098381 seconds
2015-08-26 16:37:36,860 servicerouter DEBUG updating tasks finished, took 1.35650086403 seconds
```