@flosell 2015-08-27
fixed, made the app mandatory