@cmaloney 2015-08-27
Still WIP on the review / working through it, should be more before too long. I'll comment again when I'm done (in and out of meetings)