@cmaloney 2015-08-27
Finished reviewing, only thing which is left to fix would be to make it so we fill out services first then just assume the service entry exists when looping over tasks rather than the existing sort of inverted logic currently there. Welcome to fix now or just leave it (Code works fine as is, and this is a massive improvement).

When you're happy with things rebase down to a single commit and I'll sanity-check and merge the PR