@flosell 2015-08-28
squashed commits and refactored `get_apps` along the lines you outlined.